Enable testing with python 3.2 in travis.
The previous issue was a problem building libxml2 in travis.
That library was used for regular expressions only; we are now using
python's "re" module and that has freed us from the libxml2 dependency.

pyang can now be tested with python 3.2 in travis-ci.

For python 3.3 and 3.4, we still have to change some tests, as these
new python versions randomize the dictionary order and that makes the
order of some lines in some files not deterministic.