Make pyang compatible with python 2.7, 3.2, 3.3 and 3.4.
Python 3.x: there was an issue with building libxml2 in travis.
That library was used for regular expressions only; we are now using
python's "re" module and that has freed us from the libxml2 dependency.

Python 3.3 and 3.4: besides the libxml2 issue, the dictionary keys
in python are now randomized. That made some XSD tests fail, as the
output depended on key order. The emit_xsd function now sorts the
keys explicitly.

pyang can now be used with python 2.7, 3.2, 3.3 and 3.4, with tests
in travis-ci.