Honor --max-line-length when outputting arguments (also support newline policy option)
If --max-line-length isn't specified, the default behavior should be unchanged. It it is specified, then arguments for statements that prefer single-quoting (currently 'must', 'when' and 'pattern' by default) will be quoted "smartly", which includes breaking them onto multiple lines (using the concatenation operator), with an attempt to break at reasonably sensible places.
However, if the argument contains newline characters it is ASSUMED that it has already been manually formatted using double quotes, so it's not altered.

There is also a new --yang-newline-policy option, which supports two policies: "pyang" (the default) and "verbose" (which extends various _keyword_with_trailing_newline (etc.) variables). The changes are quite heuristic and probably need some iteration in the light of experience.

Finally, the "test_yang" tests have been extended a bit. The YANG filenames can include pyang options, which makes it quite easy to test a variety of options (also, via the use of soft links) to run multiple tests on a single YANG file.