Support --transform option to specify "transform" plugins
Transforms are plugins and are similar in some respects to formats, but they don't have "emit" methods and don't output anything. They are collected into an "xforms" ("transforms") map (c.f. the "fmts" map) and can be specified alongside formats. The work of a transform is done in either the pre_validate_ctx() or post_validate_ctx() method. An additional setup_xform() mechod has been added (in the base class it's just an alias for setup_fmt()).

The idea is that a transform can "do stuff" to the statement tree (obviously very carefully!). Multiple transforms can be specified, and they will be run in the order that they were given on the command line.

I've written an "edit" transform that I will probably create pull request for, but it might not be general-purpose enough to be included in the pyang distribution.