Refactor string pattern handling to only use lxml
LXML has been previously added as a requirement, hence it can be
expected to be available for pyang.
The pattern expression, and resulting string validation are put
into a new XSDPattern class.
Original design decisions are preserved, though they don't have to:
A pattern type spec is only attached if all patterns are valid,
and value validation stops on first mismatched pattern.