Do not create closures using variables changed in loops
In short, if a function is created and refers to an outside variable,
 the last value of that variable will be used, even if set afterwards.

I.e. when a validation function is set for multiple keywords which
already had a validation function, the previous implementation would
create new functions calling an original and the added function.
Given that the original function was set as a variable in a loop,
when called, each of the new functions would call the original function
of the last keyword in the list, and not the original for its keyword.

https://github.com/PyCQA/pylint/blob/master/pylint/checkers/variables.py#L469