Add (and use) an 'ensure_hyphenated_names' instance variable
Previously ctx.opts.ctx.opts.lint_ensure_hyphenated_names was checked directly, but that makes it hard for derived classes (they would have to set the option, which is probably frowned upon). The assignment is conditional to prevent the absence of the command-line option from overriding a derived class's override.