scripts: port to entry_points.console_scripts
The current approach is not recoomended [1], and it has not worked on
Windows since the distribution switched to Python wheels. With wheels,
no code from `setup.py` runs at the installation time, which means that
platform detection logic which tried to prepare BAT files on Windows was
non-effective. The wheels are marked using the `any` tag which means
"pure Python code only", which is a correct marking, but then the actual
content of wheels would differ depending on whether they were *built* on
Windows vs. Unix.

The modern alternative is to let `setuptools` create these scripts
automatically [2]. That thing apparently puts, e.g., `pyang` into $PATH
on Windows, possibly generating all the requires wrappers and what not.
That sounds like a best thing since sliced bread, and indeed it is.

Discovered via GNPy, thanks to Stefan Melin from Telia for reporting a
test failure on Windows.

[1] https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#scripts
[2] https://setuptools.pypa.io/en/latest/userguide/quickstart.html#entry-points-and-automatic-script-creation