Correct issue with the rendering of imports
Imports of a module being processed are to be rendered as a relation from the package representing the module to a package represented each imported module.
PlantUML does not render this as desired, because the relations are defined within the package representing the given module itself causing the PlantUML keyword referencing the package in the relation to be interpreted as a new class.
This correction moves the definitions of the relations to outside of the package.