Fix dots in identifiers not being converted to underscore in PlantUML keywords
PlantUML does not accept dots in keywords, but the current UML plugin does not convert dots in keywords (as permitted for YANG identifiers by RFC 7950 section 6.2) to underscores.
Also to ensure there is no chance of a clash of keywords where say both the identifiers 'a.b' and 'a-b' exist, use 2 underscores for dots.