Move emit_child_stmt() out of emit_notify()
Whether further child statements should be rendered depends on whether cont=true when emit_child_stmt() is called and by augments it is set to false, but emit_notify() always emits child statements. So this change moves the the call to emit_child_stmt() out of emit_notify() in the same way that this is done for lists.