detect xslt_dir and schema_dir in yang2dsdl
When pyang is user-installed (i.e. into .local/), PYANG_XSLT_DIR
and PYANG_RNG_LIBDIR would have to be specified.

With this, they are looked up relative to the yang2dsl script.

When running from a git checkout, the paths are different once again,
also detected by looking for .git/ and setup.py.