JSONXSL Plugin: Add support for uint64 and int64 type for jsonxsl conversion
As part of ietf-yang-types.yang, there is a standard type of uint64 and int64 which should be defined as integer types. Currently these are not listed in the `type_class` or `union_class` definitions for the jsonxsl plugin.

See the rfc6021 definition for standard data types here:

https://datatracker.ietf.org/doc/html/rfc6021#section-3

This commit will add these data types in as unquoted / integer types similar to the other similar data types.